refactor: UserResponse 수정 및 코드 리펙토링
- UserInterest entity 폴더로 이동
- UserResponse interests String -> InterestDto로 수정
- UserService에 User 대신 UserId 넘기도록 변경
