+++ backend/src/main/java/com/cmg/comtogether/interest/dto/InterestDto.java
@@ -1,0 +1,13 @@
+package com.cmg.comtogether.interest.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor
+public class InterestDto {
+    private Long interestId;
+    private String name;
+}
